fix get_port_mappings; move test constants up
